# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Manajemen Tugas Akhir dan Skripsi Fakultas Ilmu Komputer Universitas Djuanda.",
        "title": "SIMTAS FILKOM API",
        "termsOfService": "https://filkom.unida.ac.id",
        "contact": {
            "name": "Admin SIMTAS FILKOM",
            "email": "admin@filkom.unida.ac.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/seminars": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar seminar (scope per role)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Daftar seminar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter penguji",
                        "name": "examiner_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tanggal mulai (RFC3339)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tanggal akhir (RFC3339)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Halaman (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Per halaman (default 20)",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar seminar",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/seminars/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail seminar lengkap",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Detail seminar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seminar ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail seminar",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.SeminarDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/seminars/{id}/result": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil hasil seminar lengkap (nilai akhir + breakdown per penguji)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Hasil seminar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seminar ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil seminar",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.SeminarResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/seminars/{id}/revision": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencatat catatan revisi pasca seminar (Admin + Kaprodi only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Catat revisi seminar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seminar ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Catatan revisi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.revisionNotesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catatan tersimpan",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.SeminarDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Seminar belum passed",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/seminars/{id}/schedule": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menjadwalkan/menjadwal-ulang seminar dan menunjuk penguji (Admin + Kaprodi only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Jadwalkan seminar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seminar ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Jadwal + penguji",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.ScheduleSeminarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seminar terjadwal",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.SeminarDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Jadwal tidak valid",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Konflik ruangan/penguji",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/seminars/{id}/scores": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menginput nilai seminar per komponen (Dosen Penguji ditugaskan only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Input nilai seminar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seminar ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nilai per komponen",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.SubmitScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nilai tersimpan / hasil final",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.SeminarResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Bukan penguji ditugaskan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Sudah pernah input nilai",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/theses/{thesis_id}/seminars": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengajukan seminar proposal (gate: dokumen seminar approved) (Mahasiswa pemilik only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seminar Proposal"
                ],
                "summary": "Ajukan seminar proposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thesis ID (UUID)",
                        "name": "thesis_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pengajuan dibuat",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.SeminarDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Sudah ada seminar aktif",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Gate belum terpenuhi",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.revisionNotesRequest": {
            "type": "object",
            "properties": {
                "revision_notes": {
                    "type": "string"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "usecase.ComponentScoreInput": {
            "type": "object",
            "required": [
                "component_name",
                "score"
            ],
            "properties": {
                "component_name": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                }
            }
        },
        "usecase.ComponentScoreView": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                },
                "weighted": {
                    "type": "number"
                }
            }
        },
...
        "usecase.ExaminerScoreBreakdown": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ComponentScoreView"
                    }
                },
                "examiner": {
                    "$ref": "#/definitions/usecase.UserBrief"
                },
                "examiner_score": {
                    "type": "number"
                }
            }
        },
...
        "usecase.GradingComponentView": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "usecase.ScheduleSeminarRequest": {
            "type": "object",
            "required": [
                "examiner_ids",
                "room",
                "scheduled_at"
            ],
            "properties": {
                "examiner_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "room": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                }
            }
        },
        "usecase.SeminarDetail": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "examiners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.UserBrief"
                    }
                },
                "final_score": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.SeminarScoreDetail"
                    }
                },
                "status": {
                    "type": "string"
                },
                "thesis": {
                    "$ref": "#/definitions/usecase.SeminarThesisInfo"
                }
            }
        },
        "usecase.SeminarResult": {
            "type": "object",
            "properties": {
                "examiner_scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ExaminerScoreBreakdown"
                    }
                },
                "final_score": {
                    "type": "number"
                },
                "grading_components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.GradingComponentView"
                    }
                },
                "is_complete": {
                    "type": "boolean"
                },
                "seminar_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "usecase.SeminarScoreDetail": {
            "type": "object",
            "properties": {
                "component_name": {
                    "type": "string"
                },
                "component_weight": {
                    "type": "number"
                },
                "examiner": {
                    "$ref": "#/definitions/usecase.UserBrief"
                },
                "score": {
                    "type": "number"
                }
            }
        },
        "usecase.SeminarThesisInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "student": {
                    "$ref": "#/definitions/usecase.ThesisStudentBrief"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "usecase.SubmitScoreRequest": {
            "type": "object",
            "required": [
                "scores"
            ],
            "properties": {
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ComponentScoreInput"
                    }
                }
            }
        },
...
        "usecase.ThesisStudentBrief": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                }
            }
        },
...
        "usecase.UserBrief": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}